# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bridge-LFU API",
    "version": "1.0.0",
    "description": "Bridge LFU (Licence Follow-Up) — API de suivi des licences logicielles et de l'obsolescence des équipements pour les clients de Bridge.\n\nFonctionnalités principales:\n- Authentification et gestion des utilisateurs (rôles: admin, technicien, client)\n- Gestion des clients, licences, équipements, fournisseurs, types\n- Notifications et alertes (échéances licences, obsolescence équipements)\n- Tableaux de bord et statistiques\n- Rapports exportables (JSON/CSV/PDF/Excel)\n\nConventions d’API:\n- Authentification: session via Supabase (cookies). Lancez d'abord POST /api/auth/login dans le même navigateur/onglet.\n- Pagination: paramètres `page` et `limit`, réponses standard `{ data, count, page, totalPages, hasMore }`.\n- Filtres: exposés par ressource (ex: `status`, `client_id`, `type_id`, `date_from`, `date_to`).\n- Exports: utilisez le paramètre `format` (json|csv|pdf|excel) sur les endpoints de rapports.\n\nSupport: contactez l'équipe Bridge pour l’onboarding et les accès."
  },
  "servers": [
    { "url": "/", "description": "Serveur courant" }
  ],
...
  "paths": {
...
    "/api/licenses/{licenseId}": {
      "get": { "tags": ["Licenses"], "summary": "Obtenir une licence", "parameters": [ { "in": "path", "name": "licenseId", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "responses": { "200": { "description": "Licence" }, "404": { "$ref": "#/components/responses/NotFound" } } },
      "put": { "tags": ["Licenses"], "summary": "Mettre à jour une licence", "parameters": [ { "in": "path", "name": "licenseId", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LicenseCreate" } } } }, "responses": { "200": { "description": "Licence mise à jour" }, "400": { "$ref": "#/components/responses/BadRequest" }, "404": { "$ref": "#/components/responses/NotFound" } } },
      "delete": { "tags": ["Licenses"], "summary": "Supprimer une licence", "parameters": [ { "in": "path", "name": "licenseId", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "responses": { "204": { "description": "Supprimée" }, "404": { "$ref": "#/components/responses/NotFound" } } }
    },
...
    "/api/license-types": { "get": { "tags": ["LicenseTypes"], "summary": "Lister les types de licences", "responses": { "200": { "description": "OK" } } }, "post": { "tags": ["LicenseTypes"], "summary": "Créer un type de licence", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } }, "responses": { "201": { "description": "Créé" } } } },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        }
      },
...
      "LicenseCreate": {
        "type": "object",
        "required": ["name", "expiryDate", "clientId"],
        "properties": {
          "name": { "type": "string" },
          "editor": { "type": "string" },
          "version": { "type": "string" },
          "licenseKey": { "type": "string" },
          "purchaseDate": { "type": "string", "format": "date" },
          "expiryDate": { "type": "string", "format": "date" },
          "cost": { "type": "number" },
          "clientId": { "type": "string", "format": "uuid" },
          "typeId": { "type": "string", "format": "uuid" },
          "supplierId": { "type": "string", "format": "uuid" },
          "description": { "type": "string" }
        }
      },
...
    },
    "responses": {
      "NotFound": { "description": "Ressource non trouvée", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "BadRequest": {
        "description": "Requête invalide",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    }
  }
}